Sidebar code refactor and profile picture change
